ai_agent_service/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
openai==1.3.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
pillow==10.0.0
numpy==1.24.0
requests==2.31.0